Initial setup of assignment_ansewrs_page.dart refactoring.
